Organizacion de APIs y sus respectivas validaciones

+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
 
 class ProductController extends Controller
 {
@@ -12,7 +13,7 @@ class ProductController extends Controller
 
         $this->middleware('can:DeleteProduct')->only('destroy');
         $this->middleware('can:updateProduct')->only('update');
-        $this->middleware('can:createProduct')->only('store');
+        $this->middleware('can:createProduct')->only('create');
         $this->middleware('can:searchOneProduct')->only('show');
         $this->middleware('can:showAllProducts')->only('index');
     }
@@ -37,6 +38,10 @@ class ProductController extends Controller
      */
     public function show($id){
 
+        if(! Product::find($id)){
+            return response()->json(["message" => "Product does not exist"],400);
+        }
+
         return Product::find($id);
 
     }
@@ -49,7 +54,19 @@ class ProductController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response  json(@param $product, @status: 201)
      */
-    public function store(Request $request){
+    public function create(Request $request){
+        $validator = Validator::make($request->all(), [
+            'name_product'=> 'required|string|max:255',
+            'stock' => 'required|integer',
+            'value' => 'required|numeric',
+
+        ]);
+
+            if($validator->fails()){
+
+            return response()->json($validator->errors()->toJson(), 400);
+
+            }
 
         $product = Product::create($request->all());
 
@@ -67,6 +84,19 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id){
 
+        $validator = Validator::make($request->all(), [
+            'name_product'=> 'nullable|string|max:255',
+            'stock' => 'nullable|integer',
+            'value' => 'nullable|numeric',
+
+        ]);
+
+            if($validator->fails()){
+
+            return response()->json($validator->errors()->toJson(), 400);
+
+            }
+
         $product = Product::findOrFail($id);
         $product->update($request->all());
         return response()->json($product, 200);
@@ -82,6 +112,10 @@ class ProductController extends Controller
      * @return \Illuminate\Http\Response    json(@param $product = null, @status: 204)
      */
     public function destroy($id){
+
+        if(! Product::find($id)){
+            return response()->json(["message" => "Product does not exist"],400);
+        }
 
         Product::find($id)->delete();
         return response()->json(null, 204);
